# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/patients/": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patients",
        "operationId": "get_patients_patients__get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  },
                  "title": "Response Get Patients Patients  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Patient",
        "operationId": "create_patient_patients__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patient",
        "operationId": "get_patient_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "Delete Patient",
        "operationId": "delete_patient_patients__patient_id__delete",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "patients"
        ],
        "summary": "Update Patient",
        "operationId": "update_patient_patients__patient_id__put",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Get Invoices",
        "operationId": "get_invoices_invoices__get",
        "parameters": [
          {
            "name": "show_drafts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Show Drafts"
            }
          },
          {
            "name": "only_drafts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Only Drafts"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  },
                  "title": "Response Get Invoices Invoices  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Create Invoice",
        "operationId": "create_invoice_invoices__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Get Invoice",
        "operationId": "get_invoice_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "invoices"
        ],
        "summary": "Delete Invoice",
        "operationId": "delete_invoice_invoices__invoice_id__delete",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pdf/{invoice_id}": {
      "get": {
        "tags": [
          "pdf"
        ],
        "summary": "Get Pdf Invoice",
        "operationId": "get_pdf_invoice_pdf__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Gender": {
        "type": "string",
        "enum": [
          "MALE",
          "FEMALE"
        ],
        "title": "Gender"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Invoice": {
        "properties": {
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "is_draft": {
            "type": "boolean",
            "title": "Is Draft",
            "default": true
          },
          "invoice_date": {
            "type": "string",
            "format": "date",
            "title": "Invoice Date"
          },
          "type": {
            "$ref": "#/components/schemas/InvoiceType"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "invoice_id": {
            "type": "integer",
            "title": "Invoice Id"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            },
            "type": "array",
            "title": "Items",
            "default": []
          },
          "dates": {
            "items": {
              "$ref": "#/components/schemas/InvoiceDate"
            },
            "type": "array",
            "title": "Dates"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "invoice_date",
          "type",
          "invoice_id",
          "created_at",
          "updated_at",
          "total",
          "dates",
          "patient"
        ],
        "title": "Invoice"
      },
      "InvoiceCreate": {
        "properties": {
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "is_draft": {
            "type": "boolean",
            "title": "Is Draft",
            "default": true
          },
          "invoice_date": {
            "type": "string",
            "format": "date",
            "title": "Invoice Date"
          },
          "type": {
            "$ref": "#/components/schemas/InvoiceType"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceItemCreate"
            },
            "type": "array",
            "title": "Items"
          },
          "dates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/InvoiceDateCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dates"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "invoice_date",
          "type",
          "items"
        ],
        "title": "InvoiceCreate"
      },
      "InvoiceDate": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "date_id": {
            "type": "integer",
            "title": "Date Id"
          },
          "invoice_id": {
            "type": "integer",
            "title": "Invoice Id"
          }
        },
        "type": "object",
        "required": [
          "date",
          "date_id",
          "invoice_id"
        ],
        "title": "InvoiceDate"
      },
      "InvoiceDateCreate": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "InvoiceDateCreate"
      },
      "InvoiceItem": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "invoice_id": {
            "type": "integer",
            "title": "Invoice Id"
          }
        },
        "type": "object",
        "required": [
          "description",
          "amount",
          "item_id",
          "invoice_id"
        ],
        "title": "InvoiceItem"
      },
      "InvoiceItemCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "description",
          "amount"
        ],
        "title": "InvoiceItemCreate"
      },
      "InvoiceType": {
        "type": "string",
        "enum": [
          "HP",
          "KG"
        ],
        "title": "InvoiceType"
      },
      "Patient": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "street": {
            "type": "string",
            "title": "Street"
          },
          "street_number": {
            "type": "string",
            "title": "Street Number"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "title": "Birthday"
          },
          "kilometers_to_travel": {
            "type": "number",
            "title": "Kilometers To Travel"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "label",
          "first_name",
          "last_name",
          "gender",
          "street",
          "street_number",
          "postal_code",
          "city",
          "birthday",
          "kilometers_to_travel",
          "patient_id",
          "created_at"
        ],
        "title": "Patient"
      },
      "PatientCreate": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "street": {
            "type": "string",
            "title": "Street"
          },
          "street_number": {
            "type": "string",
            "title": "Street Number"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "title": "Birthday"
          },
          "kilometers_to_travel": {
            "type": "number",
            "title": "Kilometers To Travel"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telephone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telephone"
          }
        },
        "type": "object",
        "required": [
          "label",
          "first_name",
          "last_name",
          "gender",
          "street",
          "street_number",
          "postal_code",
          "city",
          "birthday",
          "kilometers_to_travel"
        ],
        "title": "PatientCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}